// src/types/exchange/responses.ts
export interface BaseExchangeResponse {
    /** Response status indicating success or failure of the action. */
    status: "ok" | "err";
    /** Success data or error message. */
    response:
        | {
            /** Type of operation. */
            type: string;
            /** Specific data for the action. */
            data?: unknown;
        }
        | string;
}